requirements.txt
# Core Requirements
streamlit>=1.33.0
ultralytics>=8.2.0
torch>=2.2.0
torchvision>=0.17.0

# Image Processing
opencv-python-headless>=4.9.0.80
Pillow>=10.3.0
numpy>=1.26.0

# Data Handling
pandas>=2.2.0
xlsxwriter>=3.2.0
openpyxl>=3.1.2

# Additional Utilities
python-multipart>=0.0.9
tqdm>=4.66.0  # For progress bars in model processing
